Make puzzle more accessible for people with color blindness
Red and green are difficult for many people to distinguish.
Change green to orange and red to purple. It is better to
not rely on color alone to convey information. The highlighting
of the current clues uses underline and bold. Change highlight
to underline only and use bold and orange to convey a satisfied
rule. Update instructions to match new behavior.